Revert "Move Arb consts into separate file (#17100)"
This reverts commit e24e8b3878919240cc185a23e5bbc91c3b05ba5e.
