+++ core/state_transition.go
@@ -25,9 +25,9 @@ import (
 	"fmt"
 	"slices"
 
-	"github.com/erigontech/erigon-lib/chain/arb"
 	"github.com/holiman/uint256"
 
+	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/chain/params"
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/common/dbg"
@@ -233,7 +233,7 @@ func (st *StateTransition) buyGas(gasBailout bool) error {
 
 			isCancun := st.evm.ChainRules().IsCancun
 			if st.evm.ChainRules().IsArbitrum {
-				isCancun = isCancun && st.evm.Context.ArbOSVersion >= arb.ArbosVersion_20
+				isCancun = isCancun && st.evm.Context.ArbOSVersion >= chain.ArbosVersion_20
 			}
 			if isCancun {
 				maxBlobFee, overflow := new(uint256.Int).MulOverflow(st.msg.MaxFeePerBlobGas(), new(uint256.Int).SetUint64(st.msg.BlobGas()))
